requirements.txt
smbus-cffi
spidev
rpi.gpio
uncertainties
pint
oauth2client
gspread
